pyproject.toml
[tool.poetry]
name = "ankilol"
version = "0.1.0"
description = "A system for parsing and creating Anki flashcards from your existing documents"
authors = ["edmundsj <jordan.e@berkeley.edu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ankilol"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.0"
google-api-python-client = "^2.90.0"
oauth2client = "^4.1.3"
gspread = "^5.9.0"
google-auth-httplib2 = "^0.1.0"
google-auth = "^2.20.0"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
